# rpc: Minor rpc.Send optimization.