--- a/rpc/codec/conn.go
+++ b/rpc/codec/conn.go
@@ -32,7 +32,7 @@ import (
 // LZ4 benchmarks slightly faster than Snappy for pure-RPC benchmarks,
 // but slightly slower than Snappy on higher level benchmarks like the
 // ones for the Cockroach client.
-const compressionType = wire.CompressionType_SNAPPY
+const compressionType = wire.CompressionType_NONE
 
 type decompressFunc func(src []byte, uncompressedSize uint32, m proto.Message) error