@bdarnell 2016-01-29
Disabling snappy makes sense as a way to get apples-to-apples numbers for grpc, but I'd be wary of framing it as an optimization at this stage. It may slow things down for local-only testing but be a net benefit for use on real networks.