@petermattis 2016-01-30
I guess the commit message came across a bit more forcefully than intended. The benefit to disabling compression was less than ± 1%.